handle inconsistent by looking at colors
Sometimes the weight status isn't there but there is a RED weight note,
and/or the class status isn't there but there is a RED class note.  This
fix (and the "inconsistent.t" test) uses the note's color to determine
if the status should be BAD or not.

I should clean up the test a bit, but whatever.  Also I am dumping a lot
of stuff to stderr, but not saving this in db.  Perhaps I should save to
DB to show that I made an assumption where the spreadsheet was
inconsistent?  But I think at this point my assumptions are quite
safe...CT had to set the color to RED, but they forgot to type a 'B', so
I think my assumptions are fine.

At this point, checking for "ff"  if they use some other color, then
I'll miss it, but the program will crash with a croak and a note about
the line and the file.