updated tests, also store 'UNDEFINED'
tests use test::modern now.

also, storing 'UNDEFINED' as a status, and carping about that fact,
rather than skipping the inconsistent or blank records.  The problem is
that doing so will possibly break using "past months" to clean mistakes,
so I will have to put a parameter to control this behavior.